Separate 2d and 3d stuff
