+++ course6/toyapp-mehrdad/displaylib/CMakeLists.txt
@@ -4,7 +4,10 @@ add_library(displaylib
 	sdlobject.cc sdlobject.h
 	surface.cc surface.h
     types.h
-    shapes.h shapes.cc
+    shapes2d.h shapes2d.cc
+    shapes3d.h shapes3d.cc
+    vector2d.h
+    vector3d.h
 )
 set_property(TARGET displaylib PROPERTY CXX_STANDARD 14)
 target_link_libraries(displaylib SDL2)
