+++ course6/toyapp-mehrdad/displaylib/types.h
@@ -11,21 +11,6 @@ struct Pixel
 	int mY;
 };
 
-struct Vector2D
-{
-	double mX;
-	double mY;
-
-	Pixel ToPixel() { return {static_cast<int>(mX), static_cast<int>(mY)}; }
-};
-
-struct Vector3D
-{
-	double mX;
-	double mY;
-	double mZ;
-};
-
 using Degree = double;
 using Radian = double;
