+++ course6/toyapp-mehrdad/displaylib/vector2d.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "types.h"
+
+namespace Display {
+
+
+struct Vector2D
+{
+	double mX;
+	double mY;
+
+	Pixel ToPixel() { return {static_cast<int>(mX), static_cast<int>(mY)}; }
+};
+
+}
