+++ course6/toyapp-mehrdad/displaylib/vector3d.h
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace Display {
+
+struct Vector3D
+{
+	double mX;
+	double mY;
+	double mZ;
+};
+
+
+}
